hw3/requirements.txt
backports-abc==0.4
beautifulsoup4==4.4.1
bokeh==0.10.0
boto==2.38.0
bz2file==0.98
cycler==0.10.0
decorator==4.0.4
Flask==0.10.1
gensim==0.12.3
httpretty==0.8.6
ipykernel==4.1.1
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
MarkupSafe==0.23
matplotlib==2.1.0
mistune==0.7.1
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.10
nltk==3.1
notebook==4.0.6
numpy==1.13.3
pandas==0.20.3
path.py==8.1.2
patsy==0.4.1
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
Pygments==2.0.2
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.11
pyzmq==15.0.0
qtconsole==4.1.0
requests==2.8.1
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
smart-open==1.3.0
statsmodels==0.8.0
sympy==1.1.1
terminado==0.5
Theano==0.9.0
tornado==4.3
traitlets==4.0.0
Werkzeug==0.11.2
ruamel.yaml==0.0.0

